refactor: декомпозиция логики
