+++ packages/client/src/hooks/index.ts
@@ -2,3 +2,4 @@ export { useCustomIntl } from './useCustomIntl'
 export { useAppDispatch, useAppSelector } from './redux'
 export { useNavigateCustom } from './useNavigateCustom'
 export { useSnackbar } from './useSnackbar'
+export { useClickByOutZone } from './useClickByOutZone'
